backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
apscheduler==3.10.4
pytz==2024.1
pydantic==2.6.3
python-multipart==0.0.9
